pyproject.toml
[tool.poetry]
name = "opentelemetry"
version = "0.1.0"
description = ""
authors = ["asyncee <asyncee@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opentelemetry-api = "^0.12b0"
opentelemetry-sdk = "^0.12b0"
opentelemetry-exporter-jaeger = "^0.12b0"
opentelemetry-instrumentation-requests = "^0.12b0"
fastapi = "^0.61.0"
opentelemetry-instrumentation-fastapi = "^0.12b0"
uvicorn = "^0.11.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
